Adds the enable-tls config option
When the enable-tls config option is set, the callback URLs set into
the gitlab integrator relation and the Service URL set into the
Legend Studio will be HTTPS-prefixed.

This config option will not enable TLS on the Legend Application
itself, but it is meant to be used in conjunction with the
tls-secret-name config option on the Nginx Ingress Integrator Charm.